rename: `colours` -> `colour`
