+++ chess/colour.py
